requirements.txt
streamlit>=1.38
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
h5py>=3.10
tensorflow-cpu==2.15.0
